continue testing the workflow for nnunet
