+++ octopi/main.py
@@ -10,7 +10,7 @@ from octopi.entry_points.run_segment import cli as inference
 from octopi.entry_points.run_localize import cli as localize
 from octopi.entry_points.run_evaluate import cli as evaluate
 from octopi.entry_points.run_extract_mb_picks import cli as mb_extract
-from octopi.nnunet import cli as nnunet_group
+from octopi.nnunet import cli as nnunet
 
 @click.group(context_settings=cli_context)
 def routines():
@@ -26,7 +26,11 @@ routines.add_command(localize)
 routines.add_command(model_explore)
 routines.add_command(evaluate)
 routines.add_command(mb_extract)
-routines.add_command(nnunet_group)
+
+try: 
+    routines.add_command(nnunet)
+except ImportError:
+    pass
 
 @click.group(context_settings=cli_context)
 def slurm_routines():
